ProyectoFinal/app requirements.txt
flask==2.0.3
google-cloud-bigquery==3.25.0
pyarrow==14.0.0
gunicorn==22.0.0